Re-worked setup script setup.
CloudLab execution services seem to run before various other things
happen in the system, like setting up users and bashrc files and so on.
This makes making modifications to user setup (like adding environment
variables to the bashrc file) impossible. It's easier to run setup after
the machine has been fully configured, and this is what the scripts have
been re-written to do.

dist-setup.sh is used to run the setup script on each of the machines in
a cluster.