Changed LDBC parameter settings.
Increased the thread count from 1 to 32. The number of reduce tasks that
ultimately get to run is limited by the value of this parameter.
Currently setting it to 32, which will need to be changed in the future
to be dependent on the total number of cores in the system.

Also changed the yarn-site.xml settings back to exactly match those of
Arnau's configuration. Best to change only 1 variable at a time.