Fixed problem with added LAN.
Adding another LAN to connect the master to a remote blockstore resulted
in the hostname "master" being linked with the IP address for the
interface on the blockstore LAN, and another one being assigned for the
IP address connected on the client-facing LAN. This made hadoop super
confused because it was trying to connect to the hostname "master" which
was mapped to an IP address that it could not connect to (from the
slaves perspective, master mapped to 10.10.1.1, but from the master's
perspective, master mapped to 10.254.254.254).

I re-ordered the interface assignment in the profile python code to help
correct this problem, and fixed setup-all.sh in light of this problem to
generate a correct slaves file.