Assigned names to LANs and fixed ip mapping.
Since there were two LANs in the system, hostname assignments were not
reliable. For example, hostname n01 would sometimes be assigned to one
IP address and at another experiment instantiation, another. Fixed this
by assigning names to the LANs so that in /etc/hosts we can distinguish
which alias is on which lan.